minor test improvements
